# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AIDE Personal Operations Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}