# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AIDE Personal Operations Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { NoteCategory } from "./noteCategory";
import type { NoteStatus } from "./noteStatus";

export interface Note {
  id: string;
  text: string;
  category: NoteCategory;
  owner: string;
  status: NoteStatus;
  createdAt: Date;
}